removed the bulk of the code from server js and moved into its own folder. server.js is now wholly used for server related information. updated some functions as well. there is still some circular dependency so the askPrompt.js file is currently larger than it needs to be
