+++ lib/prompt.js
